Tests/Unit: remove the YoastSEO trait
On the one hand, the YoastSEO trait should by its nature be in the `doubles` directory and marked as a "mock".
On the other hand, as this trait is only used by one class, it doesn't actually make much sense to have this as a separate trait instead of integral to the test.

With this in mind, I'm removing the `YoastSEO` trait from the unit test suite in favour of adding the relevant code to the `Schema_Test` class in a similar manner as already in place in the `Twitter_Test` class.

If at some point in the future it would turn out that more tests need this exact same mock, this can be revisited, but we should code for the now, not for some hypothetical unknown future.